Fix type: ulong -> size_t and compare std::cmp_less

+++ src/linux/btop_collect.cpp
@@ -253,7 +253,7 @@ namespace Shared {
 
 			using namespace Gpu;
 			gpu_b_height_offsets.resize(gpus.size());
-			for (ulong i = 0; i < gpu_b_height_offsets.size(); ++i)
+			for (size_t i = 0; i < gpu_b_height_offsets.size(); ++i)
 				gpu_b_height_offsets[i] = gpus[i].supported_functions.gpu_utilization
 					   + gpus[i].supported_functions.pwr_usage
 					   + (gpus[i].supported_functions.mem_total or gpus[i].supported_functions.mem_used)
